refactor(time-tracker): extract constants
