+++ src/features/time-tracker/hooks/useUserSettings.ts
@@ -1,6 +1,7 @@
 import { useCallback } from "react";
 
-import { DEFAULT_WEEK_START, type UserSettings, type UseUserSettingsReturn } from "@/features/time-tracker/types";
+import { DEFAULT_WEEK_START } from "@/features/time-tracker/constants";
+import type { UserSettings, UseUserSettingsReturn } from "@/features/time-tracker/types";
 
 import { useLocalStorage } from "./useLocalStorage";
