+++ src/features/time-tracker/utils/time/weekUtilities.ts
@@ -6,8 +6,8 @@ import { eachDayOfInterval, endOfWeek, format, isSameWeek, startOfWeek } from "d
 import { zhTW } from "date-fns/locale";
 import { toZonedTime } from "date-fns-tz";
 
+import { DEFAULT_WEEK_START } from "@/features/time-tracker/constants";
 import type { WeekStartDay } from "@/features/time-tracker/types";
-import { DEFAULT_WEEK_START } from "@/features/time-tracker/types";
 
 import { getTaiwanNow, TAIWAN_TIMEZONE } from "./timeUtilities";
